Moved particles into ParticlesStorage
