+++ servers/rendering/renderer_rd/renderer_compositor_rd.cpp
@@ -155,6 +155,7 @@ void RendererCompositorRD::finalize() {
 	memdelete(scene);
 	memdelete(canvas);
 	memdelete(storage);
+	memdelete(particles_storage);
 	memdelete(light_storage);
 	memdelete(mesh_storage);
 	memdelete(material_storage);
@@ -291,6 +292,7 @@ RendererCompositorRD::RendererCompositorRD() {
 	material_storage = memnew(RendererRD::MaterialStorage);
 	mesh_storage = memnew(RendererRD::MeshStorage);
 	light_storage = memnew(RendererRD::LightStorage);
+	particles_storage = memnew(RendererRD::ParticlesStorage);
 	storage = memnew(RendererStorageRD);
 	canvas = memnew(RendererCanvasRenderRD(storage));
