+++ servers/rendering/renderer_rd/renderer_storage_rd.cpp
@@ -37,1766 +37,10 @@
 #include "renderer_compositor_rd.h"
 #include "servers/rendering/renderer_rd/storage_rd/light_storage.h"
 #include "servers/rendering/renderer_rd/storage_rd/mesh_storage.h"
+#include "servers/rendering/renderer_rd/storage_rd/particles_storage.h"
 #include "servers/rendering/renderer_rd/storage_rd/texture_storage.h"
 #include "servers/rendering/rendering_server_globals.h"
 #include "servers/rendering/shader_language.h"
-
-/* CANVAS TEXTURE */
-
-void RendererStorageRD::sampler_rd_configure_custom(float p_mipmap_bias) {
-	for (int i = 1; i < RS::CANVAS_ITEM_TEXTURE_FILTER_MAX; i++) {
-		for (int j = 1; j < RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX; j++) {
-			RD::SamplerState sampler_state;
-			switch (i) {
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_NEAREST;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_NEAREST;
-					sampler_state.max_lod = 0;
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_LINEAR;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_LINEAR;
-					sampler_state.max_lod = 0;
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST_WITH_MIPMAPS: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_NEAREST;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_NEAREST;
-					if (GLOBAL_GET("rendering/textures/default_filters/use_nearest_mipmap_filter")) {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-					} else {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_LINEAR;
-					}
-					sampler_state.lod_bias = p_mipmap_bias;
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR_WITH_MIPMAPS: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_LINEAR;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_LINEAR;
-					if (GLOBAL_GET("rendering/textures/default_filters/use_nearest_mipmap_filter")) {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-					} else {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_LINEAR;
-					}
-					sampler_state.lod_bias = p_mipmap_bias;
-
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST_WITH_MIPMAPS_ANISOTROPIC: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_NEAREST;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_NEAREST;
-					if (GLOBAL_GET("rendering/textures/default_filters/use_nearest_mipmap_filter")) {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-					} else {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_LINEAR;
-					}
-					sampler_state.lod_bias = p_mipmap_bias;
-					sampler_state.use_anisotropy = true;
-					sampler_state.anisotropy_max = 1 << int(GLOBAL_GET("rendering/textures/default_filters/anisotropic_filtering_level"));
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR_WITH_MIPMAPS_ANISOTROPIC: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_LINEAR;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_LINEAR;
-					if (GLOBAL_GET("rendering/textures/default_filters/use_nearest_mipmap_filter")) {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-					} else {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_LINEAR;
-					}
-					sampler_state.lod_bias = p_mipmap_bias;
-					sampler_state.use_anisotropy = true;
-					sampler_state.anisotropy_max = 1 << int(GLOBAL_GET("rendering/textures/default_filters/anisotropic_filtering_level"));
-
-				} break;
-				default: {
-				}
-			}
-			switch (j) {
-				case RS::CANVAS_ITEM_TEXTURE_REPEAT_DISABLED: {
-					sampler_state.repeat_u = RD::SAMPLER_REPEAT_MODE_CLAMP_TO_EDGE;
-					sampler_state.repeat_v = RD::SAMPLER_REPEAT_MODE_CLAMP_TO_EDGE;
-					sampler_state.repeat_w = RD::SAMPLER_REPEAT_MODE_CLAMP_TO_EDGE;
-
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_REPEAT_ENABLED: {
-					sampler_state.repeat_u = RD::SAMPLER_REPEAT_MODE_REPEAT;
-					sampler_state.repeat_v = RD::SAMPLER_REPEAT_MODE_REPEAT;
-					sampler_state.repeat_w = RD::SAMPLER_REPEAT_MODE_REPEAT;
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_REPEAT_MIRROR: {
-					sampler_state.repeat_u = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-					sampler_state.repeat_v = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-					sampler_state.repeat_w = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-				} break;
-				default: {
-				}
-			}
-
-			if (custom_rd_samplers[i][j].is_valid()) {
-				RD::get_singleton()->free(custom_rd_samplers[i][j]);
-			}
-
-			custom_rd_samplers[i][j] = RD::get_singleton()->sampler_create(sampler_state);
-		}
-	}
-}
-
-/* PARTICLES */
-
-RID RendererStorageRD::particles_allocate() {
-	return particles_owner.allocate_rid();
-}
-void RendererStorageRD::particles_initialize(RID p_rid) {
-	particles_owner.initialize_rid(p_rid, Particles());
-}
-
-void RendererStorageRD::particles_set_mode(RID p_particles, RS::ParticlesMode p_mode) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	if (particles->mode == p_mode) {
-		return;
-	}
-
-	_particles_free_data(particles);
-
-	particles->mode = p_mode;
-}
-
-void RendererStorageRD::particles_set_emitting(RID p_particles, bool p_emitting) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->emitting = p_emitting;
-}
-
-bool RendererStorageRD::particles_get_emitting(RID p_particles) {
-	ERR_FAIL_COND_V_MSG(RSG::threaded, false, "This function should never be used with threaded rendering, as it stalls the renderer.");
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND_V(!particles, false);
-
-	return particles->emitting;
-}
-
-void RendererStorageRD::_particles_free_data(Particles *particles) {
-	if (particles->particle_buffer.is_valid()) {
-		RD::get_singleton()->free(particles->particle_buffer);
-		particles->particle_buffer = RID();
-		RD::get_singleton()->free(particles->particle_instance_buffer);
-		particles->particle_instance_buffer = RID();
-	}
-
-	particles->userdata_count = 0;
-
-	if (particles->frame_params_buffer.is_valid()) {
-		RD::get_singleton()->free(particles->frame_params_buffer);
-		particles->frame_params_buffer = RID();
-	}
-	particles->particles_transforms_buffer_uniform_set = RID();
-
-	if (RD::get_singleton()->uniform_set_is_valid(particles->trail_bind_pose_uniform_set)) {
-		RD::get_singleton()->free(particles->trail_bind_pose_uniform_set);
-	}
-	particles->trail_bind_pose_uniform_set = RID();
-
-	if (particles->trail_bind_pose_buffer.is_valid()) {
-		RD::get_singleton()->free(particles->trail_bind_pose_buffer);
-		particles->trail_bind_pose_buffer = RID();
-	}
-	if (RD::get_singleton()->uniform_set_is_valid(particles->collision_textures_uniform_set)) {
-		RD::get_singleton()->free(particles->collision_textures_uniform_set);
-	}
-	particles->collision_textures_uniform_set = RID();
-
-	if (particles->particles_sort_buffer.is_valid()) {
-		RD::get_singleton()->free(particles->particles_sort_buffer);
-		particles->particles_sort_buffer = RID();
-		particles->particles_sort_uniform_set = RID();
-	}
-
-	if (particles->emission_buffer != nullptr) {
-		particles->emission_buffer = nullptr;
-		particles->emission_buffer_data.clear();
-		RD::get_singleton()->free(particles->emission_storage_buffer);
-		particles->emission_storage_buffer = RID();
-	}
-
-	if (RD::get_singleton()->uniform_set_is_valid(particles->particles_material_uniform_set)) {
-		//will need to be re-created
-		RD::get_singleton()->free(particles->particles_material_uniform_set);
-	}
-	particles->particles_material_uniform_set = RID();
-}
-
-void RendererStorageRD::particles_set_amount(RID p_particles, int p_amount) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	if (particles->amount == p_amount) {
-		return;
-	}
-
-	_particles_free_data(particles);
-
-	particles->amount = p_amount;
-
-	particles->prev_ticks = 0;
-	particles->phase = 0;
-	particles->prev_phase = 0;
-	particles->clear = true;
-
-	particles->dependency.changed_notify(DEPENDENCY_CHANGED_PARTICLES);
-}
-
-void RendererStorageRD::particles_set_lifetime(RID p_particles, double p_lifetime) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->lifetime = p_lifetime;
-}
-
-void RendererStorageRD::particles_set_one_shot(RID p_particles, bool p_one_shot) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->one_shot = p_one_shot;
-}
-
-void RendererStorageRD::particles_set_pre_process_time(RID p_particles, double p_time) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->pre_process_time = p_time;
-}
-void RendererStorageRD::particles_set_explosiveness_ratio(RID p_particles, real_t p_ratio) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->explosiveness = p_ratio;
-}
-void RendererStorageRD::particles_set_randomness_ratio(RID p_particles, real_t p_ratio) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->randomness = p_ratio;
-}
-
-void RendererStorageRD::particles_set_custom_aabb(RID p_particles, const AABB &p_aabb) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->custom_aabb = p_aabb;
-	particles->dependency.changed_notify(DEPENDENCY_CHANGED_AABB);
-}
-
-void RendererStorageRD::particles_set_speed_scale(RID p_particles, double p_scale) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->speed_scale = p_scale;
-}
-void RendererStorageRD::particles_set_use_local_coordinates(RID p_particles, bool p_enable) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->use_local_coords = p_enable;
-	particles->dependency.changed_notify(DEPENDENCY_CHANGED_PARTICLES);
-}
-
-void RendererStorageRD::particles_set_fixed_fps(RID p_particles, int p_fps) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->fixed_fps = p_fps;
-
-	_particles_free_data(particles);
-
-	particles->prev_ticks = 0;
-	particles->phase = 0;
-	particles->prev_phase = 0;
-	particles->clear = true;
-
-	particles->dependency.changed_notify(DEPENDENCY_CHANGED_PARTICLES);
-}
-
-void RendererStorageRD::particles_set_interpolate(RID p_particles, bool p_enable) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->interpolate = p_enable;
-}
-
-void RendererStorageRD::particles_set_fractional_delta(RID p_particles, bool p_enable) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->fractional_delta = p_enable;
-}
-
-void RendererStorageRD::particles_set_trails(RID p_particles, bool p_enable, double p_length) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	ERR_FAIL_COND(p_length < 0.1);
-	p_length = MIN(10.0, p_length);
-
-	particles->trails_enabled = p_enable;
-	particles->trail_length = p_length;
-
-	_particles_free_data(particles);
-
-	particles->prev_ticks = 0;
-	particles->phase = 0;
-	particles->prev_phase = 0;
-	particles->clear = true;
-
-	particles->dependency.changed_notify(DEPENDENCY_CHANGED_PARTICLES);
-}
-
-void RendererStorageRD::particles_set_trail_bind_poses(RID p_particles, const Vector<Transform3D> &p_bind_poses) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	if (particles->trail_bind_pose_buffer.is_valid() && particles->trail_bind_poses.size() != p_bind_poses.size()) {
-		_particles_free_data(particles);
-
-		particles->prev_ticks = 0;
-		particles->phase = 0;
-		particles->prev_phase = 0;
-		particles->clear = true;
-	}
-	particles->trail_bind_poses = p_bind_poses;
-	particles->trail_bind_poses_dirty = true;
-
-	particles->dependency.changed_notify(DEPENDENCY_CHANGED_PARTICLES);
-}
-
-void RendererStorageRD::particles_set_collision_base_size(RID p_particles, real_t p_size) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->collision_base_size = p_size;
-}
-
-void RendererStorageRD::particles_set_transform_align(RID p_particles, RS::ParticlesTransformAlign p_transform_align) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->transform_align = p_transform_align;
-}
-
-void RendererStorageRD::particles_set_process_material(RID p_particles, RID p_material) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->process_material = p_material;
-	particles->dependency.changed_notify(DEPENDENCY_CHANGED_PARTICLES); //the instance buffer may have changed
-}
-
-RID RendererStorageRD::particles_get_process_material(RID p_particles) const {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND_V(!particles, RID());
-
-	return particles->process_material;
-}
-
-void RendererStorageRD::particles_set_draw_order(RID p_particles, RS::ParticlesDrawOrder p_order) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->draw_order = p_order;
-}
-
-void RendererStorageRD::particles_set_draw_passes(RID p_particles, int p_passes) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->draw_passes.resize(p_passes);
-}
-
-void RendererStorageRD::particles_set_draw_pass_mesh(RID p_particles, int p_pass, RID p_mesh) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	ERR_FAIL_INDEX(p_pass, particles->draw_passes.size());
-	particles->draw_passes.write[p_pass] = p_mesh;
-}
-
-void RendererStorageRD::particles_restart(RID p_particles) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->restart_request = true;
-}
-
-void RendererStorageRD::_particles_allocate_emission_buffer(Particles *particles) {
-	ERR_FAIL_COND(particles->emission_buffer != nullptr);
-
-	particles->emission_buffer_data.resize(sizeof(ParticleEmissionBuffer::Data) * particles->amount + sizeof(uint32_t) * 4);
-	memset(particles->emission_buffer_data.ptrw(), 0, particles->emission_buffer_data.size());
-	particles->emission_buffer = reinterpret_cast<ParticleEmissionBuffer *>(particles->emission_buffer_data.ptrw());
-	particles->emission_buffer->particle_max = particles->amount;
-
-	particles->emission_storage_buffer = RD::get_singleton()->storage_buffer_create(particles->emission_buffer_data.size(), particles->emission_buffer_data);
-
-	if (RD::get_singleton()->uniform_set_is_valid(particles->particles_material_uniform_set)) {
-		//will need to be re-created
-		RD::get_singleton()->free(particles->particles_material_uniform_set);
-		particles->particles_material_uniform_set = RID();
-	}
-}
-
-void RendererStorageRD::particles_set_subemitter(RID p_particles, RID p_subemitter_particles) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	ERR_FAIL_COND(p_particles == p_subemitter_particles);
-
-	particles->sub_emitter = p_subemitter_particles;
-
-	if (RD::get_singleton()->uniform_set_is_valid(particles->particles_material_uniform_set)) {
-		RD::get_singleton()->free(particles->particles_material_uniform_set);
-		particles->particles_material_uniform_set = RID(); //clear and force to re create sub emitting
-	}
-}
-
-void RendererStorageRD::particles_emit(RID p_particles, const Transform3D &p_transform, const Vector3 &p_velocity, const Color &p_color, const Color &p_custom, uint32_t p_emit_flags) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	ERR_FAIL_COND(particles->amount == 0);
-
-	if (particles->emitting) {
-		particles->clear = true;
-		particles->emitting = false;
-	}
-
-	if (particles->emission_buffer == nullptr) {
-		_particles_allocate_emission_buffer(particles);
-	}
-
-	if (particles->inactive) {
-		//in case it was inactive, make active again
-		particles->inactive = false;
-		particles->inactive_time = 0;
-	}
-
-	int32_t idx = particles->emission_buffer->particle_count;
-	if (idx < particles->emission_buffer->particle_max) {
-		store_transform(p_transform, particles->emission_buffer->data[idx].xform);
-
-		particles->emission_buffer->data[idx].velocity[0] = p_velocity.x;
-		particles->emission_buffer->data[idx].velocity[1] = p_velocity.y;
-		particles->emission_buffer->data[idx].velocity[2] = p_velocity.z;
-
-		particles->emission_buffer->data[idx].custom[0] = p_custom.r;
-		particles->emission_buffer->data[idx].custom[1] = p_custom.g;
-		particles->emission_buffer->data[idx].custom[2] = p_custom.b;
-		particles->emission_buffer->data[idx].custom[3] = p_custom.a;
-
-		particles->emission_buffer->data[idx].color[0] = p_color.r;
-		particles->emission_buffer->data[idx].color[1] = p_color.g;
-		particles->emission_buffer->data[idx].color[2] = p_color.b;
-		particles->emission_buffer->data[idx].color[3] = p_color.a;
-
-		particles->emission_buffer->data[idx].flags = p_emit_flags;
-		particles->emission_buffer->particle_count++;
-	}
-}
-
-void RendererStorageRD::particles_request_process(RID p_particles) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	if (!particles->dirty) {
-		particles->dirty = true;
-		particles->update_list = particle_update_list;
-		particle_update_list = particles;
-	}
-}
-
-AABB RendererStorageRD::particles_get_current_aabb(RID p_particles) {
-	if (RSG::threaded) {
-		WARN_PRINT_ONCE("Calling this function with threaded rendering enabled stalls the renderer, use with care.");
-	}
-
-	const Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND_V(!particles, AABB());
-
-	int total_amount = particles->amount;
-	if (particles->trails_enabled && particles->trail_bind_poses.size() > 1) {
-		total_amount *= particles->trail_bind_poses.size();
-	}
-
-	Vector<uint8_t> buffer = RD::get_singleton()->buffer_get_data(particles->particle_buffer);
-	ERR_FAIL_COND_V(buffer.size() != (int)(total_amount * sizeof(ParticleData)), AABB());
-
-	Transform3D inv = particles->emission_transform.affine_inverse();
-
-	AABB aabb;
-	if (buffer.size()) {
-		bool first = true;
-
-		const uint8_t *data_ptr = (const uint8_t *)buffer.ptr();
-		uint32_t particle_data_size = sizeof(ParticleData) + sizeof(float) * particles->userdata_count;
-
-		for (int i = 0; i < total_amount; i++) {
-			const ParticleData &particle_data = *(const ParticleData *)&data_ptr[particle_data_size * i];
-			if (particle_data.active) {
-				Vector3 pos = Vector3(particle_data.xform[12], particle_data.xform[13], particle_data.xform[14]);
-				if (!particles->use_local_coords) {
-					pos = inv.xform(pos);
-				}
-				if (first) {
-					aabb.position = pos;
-					first = false;
-				} else {
-					aabb.expand_to(pos);
-				}
-			}
-		}
-	}
-
-	float longest_axis_size = 0;
-	for (int i = 0; i < particles->draw_passes.size(); i++) {
-		if (particles->draw_passes[i].is_valid()) {
-			AABB maabb = RendererRD::MeshStorage::get_singleton()->mesh_get_aabb(particles->draw_passes[i], RID());
-			longest_axis_size = MAX(maabb.get_longest_axis_size(), longest_axis_size);
-		}
-	}
-
-	aabb.grow_by(longest_axis_size);
-
-	return aabb;
-}
-
-AABB RendererStorageRD::particles_get_aabb(RID p_particles) const {
-	const Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND_V(!particles, AABB());
-
-	return particles->custom_aabb;
-}
-
-void RendererStorageRD::particles_set_emission_transform(RID p_particles, const Transform3D &p_transform) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	particles->emission_transform = p_transform;
-}
-
-int RendererStorageRD::particles_get_draw_passes(RID p_particles) const {
-	const Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND_V(!particles, 0);
-
-	return particles->draw_passes.size();
-}
-
-RID RendererStorageRD::particles_get_draw_pass_mesh(RID p_particles, int p_pass) const {
-	const Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND_V(!particles, RID());
-	ERR_FAIL_INDEX_V(p_pass, particles->draw_passes.size(), RID());
-
-	return particles->draw_passes[p_pass];
-}
-
-void RendererStorageRD::particles_add_collision(RID p_particles, RID p_particles_collision_instance) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->collisions.insert(p_particles_collision_instance);
-}
-
-void RendererStorageRD::particles_remove_collision(RID p_particles, RID p_particles_collision_instance) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->collisions.erase(p_particles_collision_instance);
-}
-
-void RendererStorageRD::particles_set_canvas_sdf_collision(RID p_particles, bool p_enable, const Transform2D &p_xform, const Rect2 &p_to_screen, RID p_texture) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-	particles->has_sdf_collision = p_enable;
-	particles->sdf_collision_transform = p_xform;
-	particles->sdf_collision_to_screen = p_to_screen;
-	particles->sdf_collision_texture = p_texture;
-}
-
-void RendererStorageRD::_particles_process(Particles *p_particles, double p_delta) {
-	RendererRD::TextureStorage *texture_storage = RendererRD::TextureStorage::get_singleton();
-	RendererRD::MaterialStorage *material_storage = RendererRD::MaterialStorage::get_singleton();
-
-	if (p_particles->particles_material_uniform_set.is_null() || !RD::get_singleton()->uniform_set_is_valid(p_particles->particles_material_uniform_set)) {
-		Vector<RD::Uniform> uniforms;
-
-		{
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-			u.binding = 0;
-			u.append_id(p_particles->frame_params_buffer);
-			uniforms.push_back(u);
-		}
-		{
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-			u.binding = 1;
-			u.append_id(p_particles->particle_buffer);
-			uniforms.push_back(u);
-		}
-
-		{
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-			u.binding = 2;
-			if (p_particles->emission_storage_buffer.is_valid()) {
-				u.append_id(p_particles->emission_storage_buffer);
-			} else {
-				u.append_id(RendererRD::MeshStorage::get_singleton()->get_default_rd_storage_buffer());
-			}
-			uniforms.push_back(u);
-		}
-		{
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-			u.binding = 3;
-			Particles *sub_emitter = particles_owner.get_or_null(p_particles->sub_emitter);
-			if (sub_emitter) {
-				if (sub_emitter->emission_buffer == nullptr) { //no emission buffer, allocate emission buffer
-					_particles_allocate_emission_buffer(sub_emitter);
-				}
-				u.append_id(sub_emitter->emission_storage_buffer);
-			} else {
-				u.append_id(RendererRD::MeshStorage::get_singleton()->get_default_rd_storage_buffer());
-			}
-			uniforms.push_back(u);
-		}
-
-		p_particles->particles_material_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, particles_shader.default_shader_rd, 1);
-	}
-
-	double new_phase = Math::fmod((double)p_particles->phase + (p_delta / p_particles->lifetime) * p_particles->speed_scale, 1.0);
-
-	//move back history (if there is any)
-	for (uint32_t i = p_particles->frame_history.size() - 1; i > 0; i--) {
-		p_particles->frame_history[i] = p_particles->frame_history[i - 1];
-	}
-	//update current frame
-	ParticlesFrameParams &frame_params = p_particles->frame_history[0];
-
-	if (p_particles->clear) {
-		p_particles->cycle_number = 0;
-		p_particles->random_seed = Math::rand();
-	} else if (new_phase < p_particles->phase) {
-		if (p_particles->one_shot) {
-			p_particles->emitting = false;
-		}
-		p_particles->cycle_number++;
-	}
-
-	frame_params.emitting = p_particles->emitting;
-	frame_params.system_phase = new_phase;
-	frame_params.prev_system_phase = p_particles->phase;
-
-	p_particles->phase = new_phase;
-
-	frame_params.time = RendererCompositorRD::singleton->get_total_time();
-	frame_params.delta = p_delta * p_particles->speed_scale;
-	frame_params.random_seed = p_particles->random_seed;
-	frame_params.explosiveness = p_particles->explosiveness;
-	frame_params.randomness = p_particles->randomness;
-
-	if (p_particles->use_local_coords) {
-		store_transform(Transform3D(), frame_params.emission_transform);
-	} else {
-		store_transform(p_particles->emission_transform, frame_params.emission_transform);
-	}
-
-	frame_params.cycle = p_particles->cycle_number;
-	frame_params.frame = p_particles->frame_counter++;
-	frame_params.pad0 = 0;
-	frame_params.pad1 = 0;
-	frame_params.pad2 = 0;
-
-	{ //collision and attractors
-
-		frame_params.collider_count = 0;
-		frame_params.attractor_count = 0;
-		frame_params.particle_size = p_particles->collision_base_size;
-
-		RID collision_3d_textures[ParticlesFrameParams::MAX_3D_TEXTURES];
-		RID collision_heightmap_texture;
-
-		Transform3D to_particles;
-		if (p_particles->use_local_coords) {
-			to_particles = p_particles->emission_transform.affine_inverse();
-		}
-
-		if (p_particles->has_sdf_collision && RD::get_singleton()->texture_is_valid(p_particles->sdf_collision_texture)) {
-			//2D collision
-
-			Transform2D xform = p_particles->sdf_collision_transform; //will use dotproduct manually so invert beforehand
-			Transform2D revert = xform.affine_inverse();
-			frame_params.collider_count = 1;
-			frame_params.colliders[0].transform[0] = xform.elements[0][0];
-			frame_params.colliders[0].transform[1] = xform.elements[0][1];
-			frame_params.colliders[0].transform[2] = 0;
-			frame_params.colliders[0].transform[3] = xform.elements[2][0];
-
-			frame_params.colliders[0].transform[4] = xform.elements[1][0];
-			frame_params.colliders[0].transform[5] = xform.elements[1][1];
-			frame_params.colliders[0].transform[6] = 0;
-			frame_params.colliders[0].transform[7] = xform.elements[2][1];
-
-			frame_params.colliders[0].transform[8] = revert.elements[0][0];
-			frame_params.colliders[0].transform[9] = revert.elements[0][1];
-			frame_params.colliders[0].transform[10] = 0;
-			frame_params.colliders[0].transform[11] = revert.elements[2][0];
-
-			frame_params.colliders[0].transform[12] = revert.elements[1][0];
-			frame_params.colliders[0].transform[13] = revert.elements[1][1];
-			frame_params.colliders[0].transform[14] = 0;
-			frame_params.colliders[0].transform[15] = revert.elements[2][1];
-
-			frame_params.colliders[0].extents[0] = p_particles->sdf_collision_to_screen.size.x;
-			frame_params.colliders[0].extents[1] = p_particles->sdf_collision_to_screen.size.y;
-			frame_params.colliders[0].extents[2] = p_particles->sdf_collision_to_screen.position.x;
-			frame_params.colliders[0].scale = p_particles->sdf_collision_to_screen.position.y;
-			frame_params.colliders[0].texture_index = 0;
-			frame_params.colliders[0].type = ParticlesFrameParams::COLLISION_TYPE_2D_SDF;
-
-			collision_heightmap_texture = p_particles->sdf_collision_texture;
-
-			//replace in all other history frames where used because parameters are no longer valid if screen moves
-			for (uint32_t i = 1; i < p_particles->frame_history.size(); i++) {
-				if (p_particles->frame_history[i].collider_count > 0 && p_particles->frame_history[i].colliders[0].type == ParticlesFrameParams::COLLISION_TYPE_2D_SDF) {
-					p_particles->frame_history[i].colliders[0] = frame_params.colliders[0];
-				}
-			}
-		}
-
-		uint32_t collision_3d_textures_used = 0;
-		for (const Set<RID>::Element *E = p_particles->collisions.front(); E; E = E->next()) {
-			ParticlesCollisionInstance *pci = particles_collision_instance_owner.get_or_null(E->get());
-			if (!pci || !pci->active) {
-				continue;
-			}
-			ParticlesCollision *pc = particles_collision_owner.get_or_null(pci->collision);
-			ERR_CONTINUE(!pc);
-
-			Transform3D to_collider = pci->transform;
-			if (p_particles->use_local_coords) {
-				to_collider = to_particles * to_collider;
-			}
-			Vector3 scale = to_collider.basis.get_scale();
-			to_collider.basis.orthonormalize();
-
-			if (pc->type <= RS::PARTICLES_COLLISION_TYPE_VECTOR_FIELD_ATTRACT) {
-				//attractor
-				if (frame_params.attractor_count >= ParticlesFrameParams::MAX_ATTRACTORS) {
-					continue;
-				}
-
-				ParticlesFrameParams::Attractor &attr = frame_params.attractors[frame_params.attractor_count];
-
-				store_transform(to_collider, attr.transform);
-				attr.strength = pc->attractor_strength;
-				attr.attenuation = pc->attractor_attenuation;
-				attr.directionality = pc->attractor_directionality;
-
-				switch (pc->type) {
-					case RS::PARTICLES_COLLISION_TYPE_SPHERE_ATTRACT: {
-						attr.type = ParticlesFrameParams::ATTRACTOR_TYPE_SPHERE;
-						float radius = pc->radius;
-						radius *= (scale.x + scale.y + scale.z) / 3.0;
-						attr.extents[0] = radius;
-						attr.extents[1] = radius;
-						attr.extents[2] = radius;
-					} break;
-					case RS::PARTICLES_COLLISION_TYPE_BOX_ATTRACT: {
-						attr.type = ParticlesFrameParams::ATTRACTOR_TYPE_BOX;
-						Vector3 extents = pc->extents * scale;
-						attr.extents[0] = extents.x;
-						attr.extents[1] = extents.y;
-						attr.extents[2] = extents.z;
-					} break;
-					case RS::PARTICLES_COLLISION_TYPE_VECTOR_FIELD_ATTRACT: {
-						if (collision_3d_textures_used >= ParticlesFrameParams::MAX_3D_TEXTURES) {
-							continue;
-						}
-						attr.type = ParticlesFrameParams::ATTRACTOR_TYPE_VECTOR_FIELD;
-						Vector3 extents = pc->extents * scale;
-						attr.extents[0] = extents.x;
-						attr.extents[1] = extents.y;
-						attr.extents[2] = extents.z;
-						attr.texture_index = collision_3d_textures_used;
-
-						collision_3d_textures[collision_3d_textures_used] = pc->field_texture;
-						collision_3d_textures_used++;
-					} break;
-					default: {
-					}
-				}
-
-				frame_params.attractor_count++;
-			} else {
-				//collider
-				if (frame_params.collider_count >= ParticlesFrameParams::MAX_COLLIDERS) {
-					continue;
-				}
-
-				ParticlesFrameParams::Collider &col = frame_params.colliders[frame_params.collider_count];
-
-				store_transform(to_collider, col.transform);
-				switch (pc->type) {
-					case RS::PARTICLES_COLLISION_TYPE_SPHERE_COLLIDE: {
-						col.type = ParticlesFrameParams::COLLISION_TYPE_SPHERE;
-						float radius = pc->radius;
-						radius *= (scale.x + scale.y + scale.z) / 3.0;
-						col.extents[0] = radius;
-						col.extents[1] = radius;
-						col.extents[2] = radius;
-					} break;
-					case RS::PARTICLES_COLLISION_TYPE_BOX_COLLIDE: {
-						col.type = ParticlesFrameParams::COLLISION_TYPE_BOX;
-						Vector3 extents = pc->extents * scale;
-						col.extents[0] = extents.x;
-						col.extents[1] = extents.y;
-						col.extents[2] = extents.z;
-					} break;
-					case RS::PARTICLES_COLLISION_TYPE_SDF_COLLIDE: {
-						if (collision_3d_textures_used >= ParticlesFrameParams::MAX_3D_TEXTURES) {
-							continue;
-						}
-						col.type = ParticlesFrameParams::COLLISION_TYPE_SDF;
-						Vector3 extents = pc->extents * scale;
-						col.extents[0] = extents.x;
-						col.extents[1] = extents.y;
-						col.extents[2] = extents.z;
-						col.texture_index = collision_3d_textures_used;
-						col.scale = (scale.x + scale.y + scale.z) * 0.333333333333; //non uniform scale non supported
-
-						collision_3d_textures[collision_3d_textures_used] = pc->field_texture;
-						collision_3d_textures_used++;
-					} break;
-					case RS::PARTICLES_COLLISION_TYPE_HEIGHTFIELD_COLLIDE: {
-						if (collision_heightmap_texture != RID()) { //already taken
-							continue;
-						}
-
-						col.type = ParticlesFrameParams::COLLISION_TYPE_HEIGHT_FIELD;
-						Vector3 extents = pc->extents * scale;
-						col.extents[0] = extents.x;
-						col.extents[1] = extents.y;
-						col.extents[2] = extents.z;
-						collision_heightmap_texture = pc->heightfield_texture;
-					} break;
-					default: {
-					}
-				}
-
-				frame_params.collider_count++;
-			}
-		}
-
-		bool different = false;
-		if (collision_3d_textures_used == p_particles->collision_3d_textures_used) {
-			for (int i = 0; i < ParticlesFrameParams::MAX_3D_TEXTURES; i++) {
-				if (p_particles->collision_3d_textures[i] != collision_3d_textures[i]) {
-					different = true;
-					break;
-				}
-			}
-		}
-
-		if (collision_heightmap_texture != p_particles->collision_heightmap_texture) {
-			different = true;
-		}
-
-		bool uniform_set_valid = RD::get_singleton()->uniform_set_is_valid(p_particles->collision_textures_uniform_set);
-
-		if (different || !uniform_set_valid) {
-			if (uniform_set_valid) {
-				RD::get_singleton()->free(p_particles->collision_textures_uniform_set);
-			}
-
-			Vector<RD::Uniform> uniforms;
-
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_TEXTURE;
-				u.binding = 0;
-				for (uint32_t i = 0; i < ParticlesFrameParams::MAX_3D_TEXTURES; i++) {
-					RID rd_tex;
-					if (i < collision_3d_textures_used) {
-						RendererRD::Texture *t = RendererRD::TextureStorage::get_singleton()->get_texture(collision_3d_textures[i]);
-						if (t && t->type == RendererRD::Texture::TYPE_3D) {
-							rd_tex = t->rd_texture;
-						}
-					}
-
-					if (rd_tex == RID()) {
-						rd_tex = texture_storage->texture_rd_get_default(RendererRD::DEFAULT_RD_TEXTURE_3D_WHITE);
-					}
-					u.append_id(rd_tex);
-				}
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_TEXTURE;
-				u.binding = 1;
-				if (collision_heightmap_texture.is_valid()) {
-					u.append_id(collision_heightmap_texture);
-				} else {
-					u.append_id(texture_storage->texture_rd_get_default(RendererRD::DEFAULT_RD_TEXTURE_BLACK));
-				}
-				uniforms.push_back(u);
-			}
-			p_particles->collision_textures_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, particles_shader.default_shader_rd, 2);
-		}
-	}
-
-	ParticlesShader::PushConstant push_constant;
-
-	int process_amount = p_particles->amount;
-
-	if (p_particles->trails_enabled && p_particles->trail_bind_poses.size() > 1) {
-		process_amount *= p_particles->trail_bind_poses.size();
-	}
-	push_constant.clear = p_particles->clear;
-	push_constant.total_particles = p_particles->amount;
-	push_constant.lifetime = p_particles->lifetime;
-	push_constant.trail_size = p_particles->trail_params.size();
-	push_constant.use_fractional_delta = p_particles->fractional_delta;
-	push_constant.sub_emitter_mode = !p_particles->emitting && p_particles->emission_buffer && (p_particles->emission_buffer->particle_count > 0 || p_particles->force_sub_emit);
-	push_constant.trail_pass = false;
-
-	p_particles->force_sub_emit = false; //reset
-
-	Particles *sub_emitter = particles_owner.get_or_null(p_particles->sub_emitter);
-
-	if (sub_emitter && sub_emitter->emission_storage_buffer.is_valid()) {
-		//	print_line("updating subemitter buffer");
-		int32_t zero[4] = { 0, sub_emitter->amount, 0, 0 };
-		RD::get_singleton()->buffer_update(sub_emitter->emission_storage_buffer, 0, sizeof(uint32_t) * 4, zero);
-		push_constant.can_emit = true;
-
-		if (sub_emitter->emitting) {
-			sub_emitter->emitting = false;
-			sub_emitter->clear = true; //will need to clear if it was emitting, sorry
-		}
-		//make sure the sub emitter processes particles too
-		sub_emitter->inactive = false;
-		sub_emitter->inactive_time = 0;
-
-		sub_emitter->force_sub_emit = true;
-
-	} else {
-		push_constant.can_emit = false;
-	}
-
-	if (p_particles->emission_buffer && p_particles->emission_buffer->particle_count) {
-		RD::get_singleton()->buffer_update(p_particles->emission_storage_buffer, 0, sizeof(uint32_t) * 4 + sizeof(ParticleEmissionBuffer::Data) * p_particles->emission_buffer->particle_count, p_particles->emission_buffer);
-		p_particles->emission_buffer->particle_count = 0;
-	}
-
-	p_particles->clear = false;
-
-	if (p_particles->trail_params.size() > 1) {
-		//fill the trail params
-		for (uint32_t i = 0; i < p_particles->trail_params.size(); i++) {
-			uint32_t src_idx = i * p_particles->frame_history.size() / p_particles->trail_params.size();
-			p_particles->trail_params[i] = p_particles->frame_history[src_idx];
-		}
-	} else {
-		p_particles->trail_params[0] = p_particles->frame_history[0];
-	}
-
-	RD::get_singleton()->buffer_update(p_particles->frame_params_buffer, 0, sizeof(ParticlesFrameParams) * p_particles->trail_params.size(), p_particles->trail_params.ptr());
-
-	ParticlesMaterialData *m = static_cast<ParticlesMaterialData *>(material_storage->material_get_data(p_particles->process_material, RendererRD::SHADER_TYPE_PARTICLES));
-	if (!m) {
-		m = static_cast<ParticlesMaterialData *>(material_storage->material_get_data(particles_shader.default_material, RendererRD::SHADER_TYPE_PARTICLES));
-	}
-
-	ERR_FAIL_COND(!m);
-
-	p_particles->has_collision_cache = m->shader_data->uses_collision;
-
-	//todo should maybe compute all particle systems together?
-	RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
-	RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, m->shader_data->pipeline);
-	RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles_shader.base_uniform_set, 0);
-	RD::get_singleton()->compute_list_bind_uniform_set(compute_list, p_particles->particles_material_uniform_set, 1);
-	RD::get_singleton()->compute_list_bind_uniform_set(compute_list, p_particles->collision_textures_uniform_set, 2);
-
-	if (m->uniform_set.is_valid() && RD::get_singleton()->uniform_set_is_valid(m->uniform_set)) {
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, m->uniform_set, 3);
-	}
-
-	RD::get_singleton()->compute_list_set_push_constant(compute_list, &push_constant, sizeof(ParticlesShader::PushConstant));
-
-	if (p_particles->trails_enabled && p_particles->trail_bind_poses.size() > 1) {
-		//trails requires two passes in order to catch particle starts
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, process_amount / p_particles->trail_bind_poses.size(), 1, 1);
-
-		RD::get_singleton()->compute_list_add_barrier(compute_list);
-
-		push_constant.trail_pass = true;
-		RD::get_singleton()->compute_list_set_push_constant(compute_list, &push_constant, sizeof(ParticlesShader::PushConstant));
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, process_amount - p_particles->amount, 1, 1);
-	} else {
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, process_amount, 1, 1);
-	}
-
-	RD::get_singleton()->compute_list_end();
-}
-
-void RendererStorageRD::particles_set_view_axis(RID p_particles, const Vector3 &p_axis, const Vector3 &p_up_axis) {
-	Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND(!particles);
-
-	if (particles->draw_order != RS::PARTICLES_DRAW_ORDER_VIEW_DEPTH && particles->transform_align != RS::PARTICLES_TRANSFORM_ALIGN_Z_BILLBOARD && particles->transform_align != RS::PARTICLES_TRANSFORM_ALIGN_Z_BILLBOARD_Y_TO_VELOCITY) {
-		return;
-	}
-
-	if (particles->particle_buffer.is_null()) {
-		return; //particles have not processed yet
-	}
-
-	bool do_sort = particles->draw_order == RS::PARTICLES_DRAW_ORDER_VIEW_DEPTH;
-
-	//copy to sort buffer
-	if (do_sort && particles->particles_sort_buffer == RID()) {
-		uint32_t size = particles->amount;
-		if (size & 1) {
-			size++; //make multiple of 16
-		}
-		size *= sizeof(float) * 2;
-		particles->particles_sort_buffer = RD::get_singleton()->storage_buffer_create(size);
-
-		{
-			Vector<RD::Uniform> uniforms;
-
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-				u.binding = 0;
-				u.append_id(particles->particles_sort_buffer);
-				uniforms.push_back(u);
-			}
-
-			particles->particles_sort_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, particles_shader.copy_shader.version_get_shader(particles_shader.copy_shader_version, ParticlesShader::COPY_MODE_FILL_SORT_BUFFER), 1);
-		}
-	}
-
-	ParticlesShader::CopyPushConstant copy_push_constant;
-
-	if (particles->trails_enabled && particles->trail_bind_poses.size() > 1) {
-		int fixed_fps = 60.0;
-		if (particles->fixed_fps > 0) {
-			fixed_fps = particles->fixed_fps;
-		}
-
-		copy_push_constant.trail_size = particles->trail_bind_poses.size();
-		copy_push_constant.trail_total = particles->frame_history.size();
-		copy_push_constant.frame_delta = 1.0 / fixed_fps;
-	} else {
-		copy_push_constant.trail_size = 1;
-		copy_push_constant.trail_total = 1;
-		copy_push_constant.frame_delta = 0.0;
-	}
-
-	copy_push_constant.order_by_lifetime = (particles->draw_order == RS::PARTICLES_DRAW_ORDER_LIFETIME || particles->draw_order == RS::PARTICLES_DRAW_ORDER_REVERSE_LIFETIME);
-	copy_push_constant.lifetime_split = MIN(particles->amount * particles->phase, particles->amount - 1);
-	copy_push_constant.lifetime_reverse = particles->draw_order == RS::PARTICLES_DRAW_ORDER_REVERSE_LIFETIME;
-
-	copy_push_constant.frame_remainder = particles->interpolate ? particles->frame_remainder : 0.0;
-	copy_push_constant.total_particles = particles->amount;
-	copy_push_constant.copy_mode_2d = false;
-
-	Vector3 axis = -p_axis; // cameras look to z negative
-
-	if (particles->use_local_coords) {
-		axis = particles->emission_transform.basis.xform_inv(axis).normalized();
-	}
-
-	copy_push_constant.sort_direction[0] = axis.x;
-	copy_push_constant.sort_direction[1] = axis.y;
-	copy_push_constant.sort_direction[2] = axis.z;
-
-	copy_push_constant.align_up[0] = p_up_axis.x;
-	copy_push_constant.align_up[1] = p_up_axis.y;
-	copy_push_constant.align_up[2] = p_up_axis.z;
-
-	copy_push_constant.align_mode = particles->transform_align;
-
-	if (do_sort) {
-		RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
-
-		RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, particles_shader.copy_pipelines[ParticlesShader::COPY_MODE_FILL_SORT_BUFFER + particles->userdata_count * ParticlesShader::COPY_MODE_MAX]);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles->particles_copy_uniform_set, 0);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles->particles_sort_uniform_set, 1);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles->trail_bind_pose_uniform_set, 2);
-		RD::get_singleton()->compute_list_set_push_constant(compute_list, &copy_push_constant, sizeof(ParticlesShader::CopyPushConstant));
-
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, particles->amount, 1, 1);
-
-		RD::get_singleton()->compute_list_end();
-		effects->sort_buffer(particles->particles_sort_uniform_set, particles->amount);
-	}
-
-	copy_push_constant.total_particles *= copy_push_constant.total_particles;
-
-	RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
-	uint32_t copy_pipeline = do_sort ? ParticlesShader::COPY_MODE_FILL_INSTANCES_WITH_SORT_BUFFER : ParticlesShader::COPY_MODE_FILL_INSTANCES;
-	copy_pipeline += particles->userdata_count * ParticlesShader::COPY_MODE_MAX;
-	copy_push_constant.copy_mode_2d = particles->mode == RS::PARTICLES_MODE_2D ? 1 : 0;
-	RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, particles_shader.copy_pipelines[copy_pipeline]);
-	RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles->particles_copy_uniform_set, 0);
-	if (do_sort) {
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles->particles_sort_uniform_set, 1);
-	}
-	RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles->trail_bind_pose_uniform_set, 2);
-
-	RD::get_singleton()->compute_list_set_push_constant(compute_list, &copy_push_constant, sizeof(ParticlesShader::CopyPushConstant));
-
-	RD::get_singleton()->compute_list_dispatch_threads(compute_list, copy_push_constant.total_particles, 1, 1);
-
-	RD::get_singleton()->compute_list_end();
-}
-
-void RendererStorageRD::_particles_update_buffers(Particles *particles) {
-	uint32_t userdata_count = 0;
-
-	const RendererRD::Material *material = RendererRD::MaterialStorage::get_singleton()->get_material(particles->process_material);
-	if (material && material->shader && material->shader->data) {
-		const ParticlesShaderData *shader_data = static_cast<const ParticlesShaderData *>(material->shader->data);
-		userdata_count = shader_data->userdata_count;
-	}
-
-	if (userdata_count != particles->userdata_count) {
-		// Mismatch userdata, re-create buffers.
-		_particles_free_data(particles);
-	}
-
-	if (particles->amount > 0 && particles->particle_buffer.is_null()) {
-		int total_amount = particles->amount;
-		if (particles->trails_enabled && particles->trail_bind_poses.size() > 1) {
-			total_amount *= particles->trail_bind_poses.size();
-		}
-
-		uint32_t xform_size = particles->mode == RS::PARTICLES_MODE_2D ? 2 : 3;
-
-		particles->particle_buffer = RD::get_singleton()->storage_buffer_create((sizeof(ParticleData) + userdata_count * sizeof(float) * 4) * total_amount);
-
-		particles->userdata_count = userdata_count;
-
-		particles->particle_instance_buffer = RD::get_singleton()->storage_buffer_create(sizeof(float) * 4 * (xform_size + 1 + 1) * total_amount);
-		//needs to clear it
-
-		{
-			Vector<RD::Uniform> uniforms;
-
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-				u.binding = 1;
-				u.append_id(particles->particle_buffer);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-				u.binding = 2;
-				u.append_id(particles->particle_instance_buffer);
-				uniforms.push_back(u);
-			}
-
-			particles->particles_copy_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, particles_shader.copy_shader.version_get_shader(particles_shader.copy_shader_version, 0), 0);
-		}
-	}
-}
-void RendererStorageRD::update_particles() {
-	while (particle_update_list) {
-		//use transform feedback to process particles
-
-		Particles *particles = particle_update_list;
-
-		//take and remove
-		particle_update_list = particles->update_list;
-		particles->update_list = nullptr;
-		particles->dirty = false;
-
-		_particles_update_buffers(particles);
-
-		if (particles->restart_request) {
-			particles->prev_ticks = 0;
-			particles->phase = 0;
-			particles->prev_phase = 0;
-			particles->clear = true;
-			particles->restart_request = false;
-		}
-
-		if (particles->inactive && !particles->emitting) {
-			//go next
-			continue;
-		}
-
-		if (particles->emitting) {
-			if (particles->inactive) {
-				//restart system from scratch
-				particles->prev_ticks = 0;
-				particles->phase = 0;
-				particles->prev_phase = 0;
-				particles->clear = true;
-			}
-			particles->inactive = false;
-			particles->inactive_time = 0;
-		} else {
-			particles->inactive_time += particles->speed_scale * RendererCompositorRD::singleton->get_frame_delta_time();
-			if (particles->inactive_time > particles->lifetime * 1.2) {
-				particles->inactive = true;
-				continue;
-			}
-		}
-
-#ifndef _MSC_VER
-#warning Should use display refresh rate for all this
-#endif
-
-		float screen_hz = 60;
-
-		int fixed_fps = 0;
-		if (particles->fixed_fps > 0) {
-			fixed_fps = particles->fixed_fps;
-		} else if (particles->trails_enabled && particles->trail_bind_poses.size() > 1) {
-			fixed_fps = screen_hz;
-		}
-		{
-			//update trails
-			int history_size = 1;
-			int trail_steps = 1;
-			if (particles->trails_enabled && particles->trail_bind_poses.size() > 1) {
-				history_size = MAX(1, int(particles->trail_length * fixed_fps));
-				trail_steps = particles->trail_bind_poses.size();
-			}
-
-			if (uint32_t(history_size) != particles->frame_history.size()) {
-				particles->frame_history.resize(history_size);
-				memset(particles->frame_history.ptr(), 0, sizeof(ParticlesFrameParams) * history_size);
-			}
-
-			if (uint32_t(trail_steps) != particles->trail_params.size() || particles->frame_params_buffer.is_null()) {
-				particles->trail_params.resize(trail_steps);
-				if (particles->frame_params_buffer.is_valid()) {
-					RD::get_singleton()->free(particles->frame_params_buffer);
-				}
-				particles->frame_params_buffer = RD::get_singleton()->storage_buffer_create(sizeof(ParticlesFrameParams) * trail_steps);
-			}
-
-			if (particles->trail_bind_poses.size() > 1 && particles->trail_bind_pose_buffer.is_null()) {
-				particles->trail_bind_pose_buffer = RD::get_singleton()->storage_buffer_create(sizeof(float) * 16 * particles->trail_bind_poses.size());
-				particles->trail_bind_poses_dirty = true;
-			}
-
-			if (particles->trail_bind_pose_uniform_set.is_null()) {
-				Vector<RD::Uniform> uniforms;
-				{
-					RD::Uniform u;
-					u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-					u.binding = 0;
-					if (particles->trail_bind_pose_buffer.is_valid()) {
-						u.append_id(particles->trail_bind_pose_buffer);
-					} else {
-						u.append_id(RendererRD::MeshStorage::get_singleton()->get_default_rd_storage_buffer());
-					}
-					uniforms.push_back(u);
-				}
-
-				particles->trail_bind_pose_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, particles_shader.copy_shader.version_get_shader(particles_shader.copy_shader_version, 0), 2);
-			}
-
-			if (particles->trail_bind_pose_buffer.is_valid() && particles->trail_bind_poses_dirty) {
-				if (particles_shader.pose_update_buffer.size() < uint32_t(particles->trail_bind_poses.size()) * 16) {
-					particles_shader.pose_update_buffer.resize(particles->trail_bind_poses.size() * 16);
-				}
-
-				for (int i = 0; i < particles->trail_bind_poses.size(); i++) {
-					store_transform(particles->trail_bind_poses[i], &particles_shader.pose_update_buffer[i * 16]);
-				}
-
-				RD::get_singleton()->buffer_update(particles->trail_bind_pose_buffer, 0, particles->trail_bind_poses.size() * 16 * sizeof(float), particles_shader.pose_update_buffer.ptr());
-			}
-		}
-
-		bool zero_time_scale = Engine::get_singleton()->get_time_scale() <= 0.0;
-
-		if (particles->clear && particles->pre_process_time > 0.0) {
-			double frame_time;
-			if (fixed_fps > 0) {
-				frame_time = 1.0 / fixed_fps;
-			} else {
-				frame_time = 1.0 / 30.0;
-			}
-
-			double todo = particles->pre_process_time;
-
-			while (todo >= 0) {
-				_particles_process(particles, frame_time);
-				todo -= frame_time;
-			}
-		}
-
-		if (fixed_fps > 0) {
-			double frame_time;
-			double decr;
-			if (zero_time_scale) {
-				frame_time = 0.0;
-				decr = 1.0 / fixed_fps;
-			} else {
-				frame_time = 1.0 / fixed_fps;
-				decr = frame_time;
-			}
-			double delta = RendererCompositorRD::singleton->get_frame_delta_time();
-			if (delta > 0.1) { //avoid recursive stalls if fps goes below 10
-				delta = 0.1;
-			} else if (delta <= 0.0) { //unlikely but..
-				delta = 0.001;
-			}
-			double todo = particles->frame_remainder + delta;
-
-			while (todo >= frame_time) {
-				_particles_process(particles, frame_time);
-				todo -= decr;
-			}
-
-			particles->frame_remainder = todo;
-
-		} else {
-			if (zero_time_scale) {
-				_particles_process(particles, 0.0);
-			} else {
-				_particles_process(particles, RendererCompositorRD::singleton->get_frame_delta_time());
-			}
-		}
-
-		//copy particles to instance buffer
-
-		if (particles->draw_order != RS::PARTICLES_DRAW_ORDER_VIEW_DEPTH && particles->transform_align != RS::PARTICLES_TRANSFORM_ALIGN_Z_BILLBOARD && particles->transform_align != RS::PARTICLES_TRANSFORM_ALIGN_Z_BILLBOARD_Y_TO_VELOCITY) {
-			//does not need view dependent operation, do copy here
-			ParticlesShader::CopyPushConstant copy_push_constant;
-
-			int total_amount = particles->amount;
-			if (particles->trails_enabled && particles->trail_bind_poses.size() > 1) {
-				total_amount *= particles->trail_bind_poses.size();
-			}
-
-			// Affect 2D only.
-			if (particles->use_local_coords) {
-				// In local mode, particle positions are calculated locally (relative to the node position)
-				// and they're also drawn locally.
-				// It works as expected, so we just pass an identity transform.
-				store_transform(Transform3D(), copy_push_constant.inv_emission_transform);
-			} else {
-				// In global mode, particle positions are calculated globally (relative to the canvas origin)
-				// but they're drawn locally.
-				// So, we need to pass the inverse of the emission transform to bring the
-				// particles to local coordinates before drawing.
-				Transform3D inv = particles->emission_transform.affine_inverse();
-				store_transform(inv, copy_push_constant.inv_emission_transform);
-			}
-
-			copy_push_constant.total_particles = total_amount;
-			copy_push_constant.frame_remainder = particles->interpolate ? particles->frame_remainder : 0.0;
-			copy_push_constant.align_mode = particles->transform_align;
-			copy_push_constant.align_up[0] = 0;
-			copy_push_constant.align_up[1] = 0;
-			copy_push_constant.align_up[2] = 0;
-
-			if (particles->trails_enabled && particles->trail_bind_poses.size() > 1) {
-				copy_push_constant.trail_size = particles->trail_bind_poses.size();
-				copy_push_constant.trail_total = particles->frame_history.size();
-				copy_push_constant.frame_delta = 1.0 / fixed_fps;
-			} else {
-				copy_push_constant.trail_size = 1;
-				copy_push_constant.trail_total = 1;
-				copy_push_constant.frame_delta = 0.0;
-			}
-
-			copy_push_constant.order_by_lifetime = (particles->draw_order == RS::PARTICLES_DRAW_ORDER_LIFETIME || particles->draw_order == RS::PARTICLES_DRAW_ORDER_REVERSE_LIFETIME);
-			copy_push_constant.lifetime_split = MIN(particles->amount * particles->phase, particles->amount - 1);
-			copy_push_constant.lifetime_reverse = particles->draw_order == RS::PARTICLES_DRAW_ORDER_REVERSE_LIFETIME;
-
-			RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
-			copy_push_constant.copy_mode_2d = particles->mode == RS::PARTICLES_MODE_2D ? 1 : 0;
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, particles_shader.copy_pipelines[ParticlesShader::COPY_MODE_FILL_INSTANCES + particles->userdata_count * ParticlesShader::COPY_MODE_MAX]);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles->particles_copy_uniform_set, 0);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, particles->trail_bind_pose_uniform_set, 2);
-			RD::get_singleton()->compute_list_set_push_constant(compute_list, &copy_push_constant, sizeof(ParticlesShader::CopyPushConstant));
-
-			RD::get_singleton()->compute_list_dispatch_threads(compute_list, total_amount, 1, 1);
-
-			RD::get_singleton()->compute_list_end();
-		}
-
-		particles->dependency.changed_notify(DEPENDENCY_CHANGED_AABB);
-	}
-}
-
-bool RendererStorageRD::particles_is_inactive(RID p_particles) const {
-	ERR_FAIL_COND_V_MSG(RSG::threaded, false, "This function should never be used with threaded rendering, as it stalls the renderer.");
-	const Particles *particles = particles_owner.get_or_null(p_particles);
-	ERR_FAIL_COND_V(!particles, false);
-	return !particles->emitting && particles->inactive;
-}
-
-/* SKY SHADER */
-
-void RendererStorageRD::ParticlesShaderData::set_code(const String &p_code) {
-	//compile
-
-	code = p_code;
-	valid = false;
-	ubo_size = 0;
-	uniforms.clear();
-	uses_collision = false;
-
-	if (code.is_empty()) {
-		return; //just invalid, but no error
-	}
-
-	ShaderCompiler::GeneratedCode gen_code;
-	ShaderCompiler::IdentifierActions actions;
-	actions.entry_point_stages["start"] = ShaderCompiler::STAGE_COMPUTE;
-	actions.entry_point_stages["process"] = ShaderCompiler::STAGE_COMPUTE;
-
-	/*
-	uses_time = false;
-
-	actions.render_mode_flags["use_half_res_pass"] = &uses_half_res;
-	actions.render_mode_flags["use_quarter_res_pass"] = &uses_quarter_res;
-
-	actions.usage_flag_pointers["TIME"] = &uses_time;
-*/
-
-	actions.usage_flag_pointers["COLLIDED"] = &uses_collision;
-
-	userdata_count = 0;
-	for (uint32_t i = 0; i < ParticlesShader::MAX_USERDATAS; i++) {
-		userdatas_used[i] = false;
-		actions.usage_flag_pointers["USERDATA" + itos(i + 1)] = &userdatas_used[i];
-	}
-
-	actions.uniforms = &uniforms;
-
-	Error err = base_singleton->particles_shader.compiler.compile(RS::SHADER_PARTICLES, code, &actions, path, gen_code);
-	ERR_FAIL_COND_MSG(err != OK, "Shader compilation failed.");
-
-	if (version.is_null()) {
-		version = base_singleton->particles_shader.shader.version_create();
-	}
-
-	for (uint32_t i = 0; i < ParticlesShader::MAX_USERDATAS; i++) {
-		if (userdatas_used[i]) {
-			userdata_count++;
-		}
-	}
-
-	base_singleton->particles_shader.shader.version_set_compute_code(version, gen_code.code, gen_code.uniforms, gen_code.stage_globals[ShaderCompiler::STAGE_COMPUTE], gen_code.defines);
-	ERR_FAIL_COND(!base_singleton->particles_shader.shader.version_is_valid(version));
-
-	ubo_size = gen_code.uniform_total_size;
-	ubo_offsets = gen_code.uniform_offsets;
-	texture_uniforms = gen_code.texture_uniforms;
-
-	//update pipelines
-
-	pipeline = RD::get_singleton()->compute_pipeline_create(base_singleton->particles_shader.shader.version_get_shader(version, 0));
-
-	valid = true;
-}
-
-void RendererStorageRD::ParticlesShaderData::set_default_texture_param(const StringName &p_name, RID p_texture, int p_index) {
-	if (!p_texture.is_valid()) {
-		if (default_texture_params.has(p_name) && default_texture_params[p_name].has(p_index)) {
-			default_texture_params[p_name].erase(p_index);
-
-			if (default_texture_params[p_name].is_empty()) {
-				default_texture_params.erase(p_name);
-			}
-		}
-	} else {
-		if (!default_texture_params.has(p_name)) {
-			default_texture_params[p_name] = Map<int, RID>();
-		}
-		default_texture_params[p_name][p_index] = p_texture;
-	}
-}
-
-void RendererStorageRD::ParticlesShaderData::get_param_list(List<PropertyInfo> *p_param_list) const {
-	Map<int, StringName> order;
-
-	for (const KeyValue<StringName, ShaderLanguage::ShaderNode::Uniform> &E : uniforms) {
-		if (E.value.scope == ShaderLanguage::ShaderNode::Uniform::SCOPE_GLOBAL || E.value.scope == ShaderLanguage::ShaderNode::Uniform::SCOPE_INSTANCE) {
-			continue;
-		}
-
-		if (E.value.texture_order >= 0) {
-			order[E.value.texture_order + 100000] = E.key;
-		} else {
-			order[E.value.order] = E.key;
-		}
-	}
-
-	for (const KeyValue<int, StringName> &E : order) {
-		PropertyInfo pi = ShaderLanguage::uniform_to_property_info(uniforms[E.value]);
-		pi.name = E.value;
-		p_param_list->push_back(pi);
-	}
-}
-
-void RendererStorageRD::ParticlesShaderData::get_instance_param_list(List<RendererMaterialStorage::InstanceShaderParam> *p_param_list) const {
-	for (const KeyValue<StringName, ShaderLanguage::ShaderNode::Uniform> &E : uniforms) {
-		if (E.value.scope != ShaderLanguage::ShaderNode::Uniform::SCOPE_INSTANCE) {
-			continue;
-		}
-
-		RendererMaterialStorage::InstanceShaderParam p;
-		p.info = ShaderLanguage::uniform_to_property_info(E.value);
-		p.info.name = E.key; //supply name
-		p.index = E.value.instance_index;
-		p.default_value = ShaderLanguage::constant_value_to_variant(E.value.default_value, E.value.type, E.value.array_size, E.value.hint);
-		p_param_list->push_back(p);
-	}
-}
-
-bool RendererStorageRD::ParticlesShaderData::is_param_texture(const StringName &p_param) const {
-	if (!uniforms.has(p_param)) {
-		return false;
-	}
-
-	return uniforms[p_param].texture_order >= 0;
-}
-
-bool RendererStorageRD::ParticlesShaderData::is_animated() const {
-	return false;
-}
-
-bool RendererStorageRD::ParticlesShaderData::casts_shadows() const {
-	return false;
-}
-
-Variant RendererStorageRD::ParticlesShaderData::get_default_parameter(const StringName &p_parameter) const {
-	if (uniforms.has(p_parameter)) {
-		ShaderLanguage::ShaderNode::Uniform uniform = uniforms[p_parameter];
-		Vector<ShaderLanguage::ConstantNode::Value> default_value = uniform.default_value;
-		return ShaderLanguage::constant_value_to_variant(default_value, uniform.type, uniform.array_size, uniform.hint);
-	}
-	return Variant();
-}
-
-RS::ShaderNativeSourceCode RendererStorageRD::ParticlesShaderData::get_native_source_code() const {
-	return base_singleton->particles_shader.shader.version_get_native_source_code(version);
-}
-
-RendererStorageRD::ParticlesShaderData::ParticlesShaderData() {
-	valid = false;
-}
-
-RendererStorageRD::ParticlesShaderData::~ParticlesShaderData() {
-	//pipeline variants will clear themselves if shader is gone
-	if (version.is_valid()) {
-		base_singleton->particles_shader.shader.version_free(version);
-	}
-}
-
-RendererRD::ShaderData *RendererStorageRD::_create_particles_shader_func() {
-	ParticlesShaderData *shader_data = memnew(ParticlesShaderData);
-	return shader_data;
-}
-
-bool RendererStorageRD::ParticlesMaterialData::update_parameters(const Map<StringName, Variant> &p_parameters, bool p_uniform_dirty, bool p_textures_dirty) {
-	return update_parameters_uniform_set(p_parameters, p_uniform_dirty, p_textures_dirty, shader_data->uniforms, shader_data->ubo_offsets.ptr(), shader_data->texture_uniforms, shader_data->default_texture_params, shader_data->ubo_size, uniform_set, base_singleton->particles_shader.shader.version_get_shader(shader_data->version, 0), 3);
-}
-
-RendererStorageRD::ParticlesMaterialData::~ParticlesMaterialData() {
-	free_parameters_uniform_set(uniform_set);
-}
-
-RendererRD::MaterialData *RendererStorageRD::_create_particles_material_func(ParticlesShaderData *p_shader) {
-	ParticlesMaterialData *material_data = memnew(ParticlesMaterialData);
-	material_data->shader_data = p_shader;
-	//update will happen later anyway so do nothing.
-	return material_data;
-}
-////////
-
-/* PARTICLES COLLISION API */
-
-RID RendererStorageRD::particles_collision_allocate() {
-	return particles_collision_owner.allocate_rid();
-}
-void RendererStorageRD::particles_collision_initialize(RID p_rid) {
-	particles_collision_owner.initialize_rid(p_rid, ParticlesCollision());
-}
-
-RID RendererStorageRD::particles_collision_get_heightfield_framebuffer(RID p_particles_collision) const {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND_V(!particles_collision, RID());
-	ERR_FAIL_COND_V(particles_collision->type != RS::PARTICLES_COLLISION_TYPE_HEIGHTFIELD_COLLIDE, RID());
-
-	if (particles_collision->heightfield_texture == RID()) {
-		//create
-		int resolutions[RS::PARTICLES_COLLISION_HEIGHTFIELD_RESOLUTION_MAX] = { 256, 512, 1024, 2048, 4096, 8192 };
-		Size2i size;
-		if (particles_collision->extents.x > particles_collision->extents.z) {
-			size.x = resolutions[particles_collision->heightfield_resolution];
-			size.y = int32_t(particles_collision->extents.z / particles_collision->extents.x * size.x);
-		} else {
-			size.y = resolutions[particles_collision->heightfield_resolution];
-			size.x = int32_t(particles_collision->extents.x / particles_collision->extents.z * size.y);
-		}
-
-		RD::TextureFormat tf;
-		tf.format = RD::DATA_FORMAT_D32_SFLOAT;
-		tf.width = size.x;
-		tf.height = size.y;
-		tf.texture_type = RD::TEXTURE_TYPE_2D;
-		tf.usage_bits = RD::TEXTURE_USAGE_SAMPLING_BIT | RD::TEXTURE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT;
-
-		particles_collision->heightfield_texture = RD::get_singleton()->texture_create(tf, RD::TextureView());
-
-		Vector<RID> fb_tex;
-		fb_tex.push_back(particles_collision->heightfield_texture);
-		particles_collision->heightfield_fb = RD::get_singleton()->framebuffer_create(fb_tex);
-		particles_collision->heightfield_fb_size = size;
-	}
-
-	return particles_collision->heightfield_fb;
-}
-
-void RendererStorageRD::particles_collision_set_collision_type(RID p_particles_collision, RS::ParticlesCollisionType p_type) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-
-	if (p_type == particles_collision->type) {
-		return;
-	}
-
-	if (particles_collision->heightfield_texture.is_valid()) {
-		RD::get_singleton()->free(particles_collision->heightfield_texture);
-		particles_collision->heightfield_texture = RID();
-	}
-	particles_collision->type = p_type;
-	particles_collision->dependency.changed_notify(DEPENDENCY_CHANGED_AABB);
-}
-
-void RendererStorageRD::particles_collision_set_cull_mask(RID p_particles_collision, uint32_t p_cull_mask) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-	particles_collision->cull_mask = p_cull_mask;
-}
-
-void RendererStorageRD::particles_collision_set_sphere_radius(RID p_particles_collision, real_t p_radius) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-
-	particles_collision->radius = p_radius;
-	particles_collision->dependency.changed_notify(DEPENDENCY_CHANGED_AABB);
-}
-
-void RendererStorageRD::particles_collision_set_box_extents(RID p_particles_collision, const Vector3 &p_extents) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-
-	particles_collision->extents = p_extents;
-	particles_collision->dependency.changed_notify(DEPENDENCY_CHANGED_AABB);
-}
-
-void RendererStorageRD::particles_collision_set_attractor_strength(RID p_particles_collision, real_t p_strength) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-
-	particles_collision->attractor_strength = p_strength;
-}
-
-void RendererStorageRD::particles_collision_set_attractor_directionality(RID p_particles_collision, real_t p_directionality) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-
-	particles_collision->attractor_directionality = p_directionality;
-}
-
-void RendererStorageRD::particles_collision_set_attractor_attenuation(RID p_particles_collision, real_t p_curve) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-
-	particles_collision->attractor_attenuation = p_curve;
-}
-
-void RendererStorageRD::particles_collision_set_field_texture(RID p_particles_collision, RID p_texture) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-
-	particles_collision->field_texture = p_texture;
-}
-
-void RendererStorageRD::particles_collision_height_field_update(RID p_particles_collision) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-	particles_collision->dependency.changed_notify(DEPENDENCY_CHANGED_AABB);
-}
-
-void RendererStorageRD::particles_collision_set_height_field_resolution(RID p_particles_collision, RS::ParticlesCollisionHeightfieldResolution p_resolution) {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND(!particles_collision);
-	ERR_FAIL_INDEX(p_resolution, RS::PARTICLES_COLLISION_HEIGHTFIELD_RESOLUTION_MAX);
-
-	if (particles_collision->heightfield_resolution == p_resolution) {
-		return;
-	}
-
-	particles_collision->heightfield_resolution = p_resolution;
-
-	if (particles_collision->heightfield_texture.is_valid()) {
-		RD::get_singleton()->free(particles_collision->heightfield_texture);
-		particles_collision->heightfield_texture = RID();
-	}
-}
-
-AABB RendererStorageRD::particles_collision_get_aabb(RID p_particles_collision) const {
-	ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND_V(!particles_collision, AABB());
-
-	switch (particles_collision->type) {
-		case RS::PARTICLES_COLLISION_TYPE_SPHERE_ATTRACT:
-		case RS::PARTICLES_COLLISION_TYPE_SPHERE_COLLIDE: {
-			AABB aabb;
-			aabb.position = -Vector3(1, 1, 1) * particles_collision->radius;
-			aabb.size = Vector3(2, 2, 2) * particles_collision->radius;
-			return aabb;
-		}
-		default: {
-			AABB aabb;
-			aabb.position = -particles_collision->extents;
-			aabb.size = particles_collision->extents * 2;
-			return aabb;
-		}
-	}
-
-	return AABB();
-}
-
-Vector3 RendererStorageRD::particles_collision_get_extents(RID p_particles_collision) const {
-	const ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND_V(!particles_collision, Vector3());
-	return particles_collision->extents;
-}
-
-bool RendererStorageRD::particles_collision_is_heightfield(RID p_particles_collision) const {
-	const ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_particles_collision);
-	ERR_FAIL_COND_V(!particles_collision, false);
-	return particles_collision->type == RS::PARTICLES_COLLISION_TYPE_HEIGHTFIELD_COLLIDE;
-}
-
-RID RendererStorageRD::particles_collision_instance_create(RID p_collision) {
-	ParticlesCollisionInstance pci;
-	pci.collision = p_collision;
-	return particles_collision_instance_owner.make_rid(pci);
-}
-void RendererStorageRD::particles_collision_instance_set_transform(RID p_collision_instance, const Transform3D &p_transform) {
-	ParticlesCollisionInstance *pci = particles_collision_instance_owner.get_or_null(p_collision_instance);
-	ERR_FAIL_COND(!pci);
-	pci->transform = p_transform;
-}
-void RendererStorageRD::particles_collision_instance_set_active(RID p_collision_instance, bool p_active) {
-	ParticlesCollisionInstance *pci = particles_collision_instance_owner.get_or_null(p_collision_instance);
-	ERR_FAIL_COND(!pci);
-	pci->active = p_active;
-}
 
 /* FOG VOLUMES */
 
@@ -2257,6 +501,8 @@ RID RendererStorageRD::voxel_gi_get_sdf_texture(RID p_voxel_gi) {
 	return voxel_gi->sdf_texture;
 }
 
+/* misc */
+
 void RendererStorageRD::base_update_dependency(RID p_base, DependencyTracker *p_instance) {
 	if (RendererRD::MeshStorage::get_singleton()->owns_mesh(p_base)) {
 		RendererRD::Mesh *mesh = RendererRD::MeshStorage::get_singleton()->get_mesh(p_base);
@@ -2282,11 +528,11 @@ void RendererStorageRD::base_update_dependency(RID p_base, DependencyTracker *p_
 	} else if (RendererRD::LightStorage::get_singleton()->owns_light(p_base)) {
 		RendererRD::Light *l = RendererRD::LightStorage::get_singleton()->get_light(p_base);
 		p_instance->update_dependency(&l->dependency);
-	} else if (particles_owner.owns(p_base)) {
-		Particles *p = particles_owner.get_or_null(p_base);
+	} else if (RendererRD::ParticlesStorage::get_singleton()->owns_particles(p_base)) {
+		RendererRD::Particles *p = RendererRD::ParticlesStorage::get_singleton()->get_particles(p_base);
 		p_instance->update_dependency(&p->dependency);
-	} else if (particles_collision_owner.owns(p_base)) {
-		ParticlesCollision *pc = particles_collision_owner.get_or_null(p_base);
+	} else if (RendererRD::ParticlesStorage::get_singleton()->owns_particles_collision(p_base)) {
+		RendererRD::ParticlesCollision *pc = RendererRD::ParticlesStorage::get_singleton()->get_particles_collision(p_base);
 		p_instance->update_dependency(&pc->dependency);
 	} else if (fog_volume_owner.owns(p_base)) {
 		FogVolume *fv = fog_volume_owner.get_or_null(p_base);
@@ -2319,10 +565,10 @@ RS::InstanceType RendererStorageRD::get_base_type(RID p_rid) const {
 	if (RendererRD::LightStorage::get_singleton()->owns_lightmap(p_rid)) {
 		return RS::INSTANCE_LIGHTMAP;
 	}
-	if (particles_owner.owns(p_rid)) {
+	if (RendererRD::ParticlesStorage::get_singleton()->owns_particles(p_rid)) {
 		return RS::INSTANCE_PARTICLES;
 	}
-	if (particles_collision_owner.owns(p_rid)) {
+	if (RendererRD::ParticlesStorage::get_singleton()->owns_particles_collision(p_rid)) {
 		return RS::INSTANCE_PARTICLES_COLLISION;
 	}
 	if (fog_volume_owner.owns(p_rid)) {
@@ -2393,26 +639,16 @@ bool RendererStorageRD::free(RID p_rid) {
 		RendererRD::LightStorage::get_singleton()->lightmap_free(p_rid);
 	} else if (RendererRD::LightStorage::get_singleton()->owns_light(p_rid)) {
 		RendererRD::LightStorage::get_singleton()->light_free(p_rid);
-	} else if (particles_owner.owns(p_rid)) {
-		update_particles();
-		Particles *particles = particles_owner.get_or_null(p_rid);
-		particles->dependency.deleted_notify(p_rid);
-		_particles_free_data(particles);
-		particles_owner.free(p_rid);
-	} else if (particles_collision_owner.owns(p_rid)) {
-		ParticlesCollision *particles_collision = particles_collision_owner.get_or_null(p_rid);
-
-		if (particles_collision->heightfield_texture.is_valid()) {
-			RD::get_singleton()->free(particles_collision->heightfield_texture);
-		}
-		particles_collision->dependency.deleted_notify(p_rid);
-		particles_collision_owner.free(p_rid);
+	} else if (RendererRD::ParticlesStorage::get_singleton()->owns_particles(p_rid)) {
+		RendererRD::ParticlesStorage::get_singleton()->particles_free(p_rid);
+	} else if (RendererRD::ParticlesStorage::get_singleton()->owns_particles_collision(p_rid)) {
+		RendererRD::ParticlesStorage::get_singleton()->particles_collision_free(p_rid);
 	} else if (visibility_notifier_owner.owns(p_rid)) {
 		VisibilityNotifier *vn = visibility_notifier_owner.get_or_null(p_rid);
 		vn->dependency.deleted_notify(p_rid);
 		visibility_notifier_owner.free(p_rid);
-	} else if (particles_collision_instance_owner.owns(p_rid)) {
-		particles_collision_instance_owner.free(p_rid);
+	} else if (RendererRD::ParticlesStorage::get_singleton()->owns_particles_collision_instance(p_rid)) {
+		RendererRD::ParticlesStorage::get_singleton()->particles_collision_instance_free(p_rid);
 	} else if (fog_volume_owner.owns(p_rid)) {
 		FogVolume *fog_volume = fog_volume_owner.get_or_null(p_rid);
 		fog_volume->dependency.deleted_notify(p_rid);
@@ -2495,272 +731,9 @@ RendererStorageRD *RendererStorageRD::base_singleton = nullptr;
 
 RendererStorageRD::RendererStorageRD() {
 	base_singleton = this;
-
-	RendererRD::MaterialStorage *material_storage = RendererRD::MaterialStorage::get_singleton();
-
-	//default samplers
-	for (int i = 1; i < RS::CANVAS_ITEM_TEXTURE_FILTER_MAX; i++) {
-		for (int j = 1; j < RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX; j++) {
-			RD::SamplerState sampler_state;
-			switch (i) {
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_NEAREST;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_NEAREST;
-					sampler_state.max_lod = 0;
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_LINEAR;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_LINEAR;
-					sampler_state.max_lod = 0;
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST_WITH_MIPMAPS: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_NEAREST;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_NEAREST;
-					if (GLOBAL_GET("rendering/textures/default_filters/use_nearest_mipmap_filter")) {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-					} else {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_LINEAR;
-					}
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR_WITH_MIPMAPS: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_LINEAR;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_LINEAR;
-					if (GLOBAL_GET("rendering/textures/default_filters/use_nearest_mipmap_filter")) {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-					} else {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_LINEAR;
-					}
-
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST_WITH_MIPMAPS_ANISOTROPIC: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_NEAREST;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_NEAREST;
-					if (GLOBAL_GET("rendering/textures/default_filters/use_nearest_mipmap_filter")) {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-					} else {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_LINEAR;
-					}
-					sampler_state.use_anisotropy = true;
-					sampler_state.anisotropy_max = 1 << int(GLOBAL_GET("rendering/textures/default_filters/anisotropic_filtering_level"));
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR_WITH_MIPMAPS_ANISOTROPIC: {
-					sampler_state.mag_filter = RD::SAMPLER_FILTER_LINEAR;
-					sampler_state.min_filter = RD::SAMPLER_FILTER_LINEAR;
-					if (GLOBAL_GET("rendering/textures/default_filters/use_nearest_mipmap_filter")) {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-					} else {
-						sampler_state.mip_filter = RD::SAMPLER_FILTER_LINEAR;
-					}
-					sampler_state.use_anisotropy = true;
-					sampler_state.anisotropy_max = 1 << int(GLOBAL_GET("rendering/textures/default_filters/anisotropic_filtering_level"));
-
-				} break;
-				default: {
-				}
-			}
-			switch (j) {
-				case RS::CANVAS_ITEM_TEXTURE_REPEAT_DISABLED: {
-					sampler_state.repeat_u = RD::SAMPLER_REPEAT_MODE_CLAMP_TO_EDGE;
-					sampler_state.repeat_v = RD::SAMPLER_REPEAT_MODE_CLAMP_TO_EDGE;
-					sampler_state.repeat_w = RD::SAMPLER_REPEAT_MODE_CLAMP_TO_EDGE;
-
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_REPEAT_ENABLED: {
-					sampler_state.repeat_u = RD::SAMPLER_REPEAT_MODE_REPEAT;
-					sampler_state.repeat_v = RD::SAMPLER_REPEAT_MODE_REPEAT;
-					sampler_state.repeat_w = RD::SAMPLER_REPEAT_MODE_REPEAT;
-				} break;
-				case RS::CANVAS_ITEM_TEXTURE_REPEAT_MIRROR: {
-					sampler_state.repeat_u = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-					sampler_state.repeat_v = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-					sampler_state.repeat_w = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-				} break;
-				default: {
-				}
-			}
-
-			default_rd_samplers[i][j] = RD::get_singleton()->sampler_create(sampler_state);
-		}
-	}
-
-	//custom sampler
-	sampler_rd_configure_custom(0.0f);
-
-	/* Particles */
-
-	{
-		// Initialize particles
-		Vector<String> particles_modes;
-		particles_modes.push_back("");
-		particles_shader.shader.initialize(particles_modes, String());
-	}
-	RendererRD::MaterialStorage::get_singleton()->shader_set_data_request_function(RendererRD::SHADER_TYPE_PARTICLES, _create_particles_shader_funcs);
-	RendererRD::MaterialStorage::get_singleton()->material_set_data_request_function(RendererRD::SHADER_TYPE_PARTICLES, _create_particles_material_funcs);
-
-	{
-		ShaderCompiler::DefaultIdentifierActions actions;
-
-		actions.renames["COLOR"] = "PARTICLE.color";
-		actions.renames["VELOCITY"] = "PARTICLE.velocity";
-		//actions.renames["MASS"] = "mass"; ?
-		actions.renames["ACTIVE"] = "particle_active";
-		actions.renames["RESTART"] = "restart";
-		actions.renames["CUSTOM"] = "PARTICLE.custom";
-		for (int i = 0; i < ParticlesShader::MAX_USERDATAS; i++) {
-			String udname = "USERDATA" + itos(i + 1);
-			actions.renames[udname] = "PARTICLE.userdata" + itos(i + 1);
-			actions.usage_defines[udname] = "#define USERDATA" + itos(i + 1) + "_USED\n";
-		}
-		actions.renames["TRANSFORM"] = "PARTICLE.xform";
-		actions.renames["TIME"] = "frame_history.data[0].time";
-		actions.renames["PI"] = _MKSTR(Math_PI);
-		actions.renames["TAU"] = _MKSTR(Math_TAU);
-		actions.renames["E"] = _MKSTR(Math_E);
-		actions.renames["LIFETIME"] = "params.lifetime";
-		actions.renames["DELTA"] = "local_delta";
-		actions.renames["NUMBER"] = "particle_number";
-		actions.renames["INDEX"] = "index";
-		//actions.renames["GRAVITY"] = "current_gravity";
-		actions.renames["EMISSION_TRANSFORM"] = "FRAME.emission_transform";
-		actions.renames["RANDOM_SEED"] = "FRAME.random_seed";
-		actions.renames["FLAG_EMIT_POSITION"] = "EMISSION_FLAG_HAS_POSITION";
-		actions.renames["FLAG_EMIT_ROT_SCALE"] = "EMISSION_FLAG_HAS_ROTATION_SCALE";
-		actions.renames["FLAG_EMIT_VELOCITY"] = "EMISSION_FLAG_HAS_VELOCITY";
-		actions.renames["FLAG_EMIT_COLOR"] = "EMISSION_FLAG_HAS_COLOR";
-		actions.renames["FLAG_EMIT_CUSTOM"] = "EMISSION_FLAG_HAS_CUSTOM";
-		actions.renames["RESTART_POSITION"] = "restart_position";
-		actions.renames["RESTART_ROT_SCALE"] = "restart_rotation_scale";
-		actions.renames["RESTART_VELOCITY"] = "restart_velocity";
-		actions.renames["RESTART_COLOR"] = "restart_color";
-		actions.renames["RESTART_CUSTOM"] = "restart_custom";
-		actions.renames["emit_subparticle"] = "emit_subparticle";
-		actions.renames["COLLIDED"] = "collided";
-		actions.renames["COLLISION_NORMAL"] = "collision_normal";
-		actions.renames["COLLISION_DEPTH"] = "collision_depth";
-		actions.renames["ATTRACTOR_FORCE"] = "attractor_force";
-
-		actions.render_mode_defines["disable_force"] = "#define DISABLE_FORCE\n";
-		actions.render_mode_defines["disable_velocity"] = "#define DISABLE_VELOCITY\n";
-		actions.render_mode_defines["keep_data"] = "#define ENABLE_KEEP_DATA\n";
-		actions.render_mode_defines["collision_use_scale"] = "#define USE_COLLISON_SCALE\n";
-
-		actions.sampler_array_name = "material_samplers";
-		actions.base_texture_binding_index = 1;
-		actions.texture_layout_set = 3;
-		actions.base_uniform_string = "material.";
-		actions.base_varying_index = 10;
-
-		actions.default_filter = ShaderLanguage::FILTER_LINEAR_MIPMAP;
-		actions.default_repeat = ShaderLanguage::REPEAT_ENABLE;
-		actions.global_buffer_array_variable = "global_variables.data";
-
-		particles_shader.compiler.initialize(actions);
-	}
-
-	{
-		// default material and shader for particles shader
-		particles_shader.default_shader = material_storage->shader_allocate();
-		material_storage->shader_initialize(particles_shader.default_shader);
-		material_storage->shader_set_code(particles_shader.default_shader, R"(
-// Default particles shader.
-
-shader_type particles;
-
-void process() {
-	COLOR = vec4(1.0);
-}
-)");
-		particles_shader.default_material = material_storage->material_allocate();
-		material_storage->material_initialize(particles_shader.default_material);
-		material_storage->material_set_shader(particles_shader.default_material, particles_shader.default_shader);
-
-		ParticlesMaterialData *md = static_cast<ParticlesMaterialData *>(material_storage->material_get_data(particles_shader.default_material, RendererRD::SHADER_TYPE_PARTICLES));
-		particles_shader.default_shader_rd = particles_shader.shader.version_get_shader(md->shader_data->version, 0);
-
-		Vector<RD::Uniform> uniforms;
-
-		{
-			Vector<RID> ids;
-			ids.resize(12);
-			RID *ids_ptr = ids.ptrw();
-			ids_ptr[0] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST, RS::CANVAS_ITEM_TEXTURE_REPEAT_DISABLED);
-			ids_ptr[1] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR, RS::CANVAS_ITEM_TEXTURE_REPEAT_DISABLED);
-			ids_ptr[2] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST_WITH_MIPMAPS, RS::CANVAS_ITEM_TEXTURE_REPEAT_DISABLED);
-			ids_ptr[3] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR_WITH_MIPMAPS, RS::CANVAS_ITEM_TEXTURE_REPEAT_DISABLED);
-			ids_ptr[4] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST_WITH_MIPMAPS_ANISOTROPIC, RS::CANVAS_ITEM_TEXTURE_REPEAT_DISABLED);
-			ids_ptr[5] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR_WITH_MIPMAPS_ANISOTROPIC, RS::CANVAS_ITEM_TEXTURE_REPEAT_DISABLED);
-			ids_ptr[6] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST, RS::CANVAS_ITEM_TEXTURE_REPEAT_ENABLED);
-			ids_ptr[7] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR, RS::CANVAS_ITEM_TEXTURE_REPEAT_ENABLED);
-			ids_ptr[8] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST_WITH_MIPMAPS, RS::CANVAS_ITEM_TEXTURE_REPEAT_ENABLED);
-			ids_ptr[9] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR_WITH_MIPMAPS, RS::CANVAS_ITEM_TEXTURE_REPEAT_ENABLED);
-			ids_ptr[10] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_NEAREST_WITH_MIPMAPS_ANISOTROPIC, RS::CANVAS_ITEM_TEXTURE_REPEAT_ENABLED);
-			ids_ptr[11] = sampler_rd_get_default(RS::CANVAS_ITEM_TEXTURE_FILTER_LINEAR_WITH_MIPMAPS_ANISOTROPIC, RS::CANVAS_ITEM_TEXTURE_REPEAT_ENABLED);
-
-			RD::Uniform u(RD::UNIFORM_TYPE_SAMPLER, 1, ids);
-			uniforms.push_back(u);
-		}
-
-		{
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-			u.binding = 2;
-			u.append_id(material_storage->global_variables_get_storage_buffer());
-			uniforms.push_back(u);
-		}
-
-		particles_shader.base_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, particles_shader.default_shader_rd, 0);
-	}
-
-	{
-		Vector<String> copy_modes;
-		for (int i = 0; i <= ParticlesShader::MAX_USERDATAS; i++) {
-			if (i == 0) {
-				copy_modes.push_back("\n#define MODE_FILL_INSTANCES\n");
-				copy_modes.push_back("\n#define MODE_FILL_SORT_BUFFER\n#define USE_SORT_BUFFER\n");
-				copy_modes.push_back("\n#define MODE_FILL_INSTANCES\n#define USE_SORT_BUFFER\n");
-			} else {
-				copy_modes.push_back("\n#define MODE_FILL_INSTANCES\n#define USERDATA_COUNT " + itos(i) + "\n");
-				copy_modes.push_back("\n#define MODE_FILL_SORT_BUFFER\n#define USE_SORT_BUFFER\n#define USERDATA_COUNT " + itos(i) + "\n");
-				copy_modes.push_back("\n#define MODE_FILL_INSTANCES\n#define USE_SORT_BUFFER\n#define USERDATA_COUNT " + itos(i) + "\n");
-			}
-		}
-
-		particles_shader.copy_shader.initialize(copy_modes);
-
-		particles_shader.copy_shader_version = particles_shader.copy_shader.version_create();
-
-		for (int i = 0; i <= ParticlesShader::MAX_USERDATAS; i++) {
-			for (int j = 0; j < ParticlesShader::COPY_MODE_MAX; j++) {
-				particles_shader.copy_pipelines[i * ParticlesShader::COPY_MODE_MAX + j] = RD::get_singleton()->compute_pipeline_create(particles_shader.copy_shader.version_get_shader(particles_shader.copy_shader_version, i * ParticlesShader::COPY_MODE_MAX + j));
-			}
-		}
-	}
 }
 
 RendererStorageRD::~RendererStorageRD() {
-	RendererRD::MaterialStorage *material_storage = RendererRD::MaterialStorage::get_singleton();
-
-	//def samplers
-	for (int i = 1; i < RS::CANVAS_ITEM_TEXTURE_FILTER_MAX; i++) {
-		for (int j = 1; j < RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX; j++) {
-			RD::get_singleton()->free(default_rd_samplers[i][j]);
-		}
-	}
-
-	//custom samplers
-	for (int i = 1; i < RS::CANVAS_ITEM_TEXTURE_FILTER_MAX; i++) {
-		for (int j = 0; j < RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX; j++) {
-			if (custom_rd_samplers[i][j].is_valid()) {
-				RD::get_singleton()->free(custom_rd_samplers[i][j]);
-			}
-		}
-	}
-
-	particles_shader.copy_shader.version_free(particles_shader.copy_shader_version);
-
-	material_storage->material_free(particles_shader.default_material);
-	material_storage->shader_free(particles_shader.default_shader);
-
 	if (effects) {
 		memdelete(effects);
 		effects = nullptr;
