+++ servers/rendering/renderer_rd/storage_rd/material_storage.h
@@ -195,6 +195,11 @@ private:
 	friend struct MaterialData;
 	static MaterialStorage *singleton;
 
+	/* Samplers */
+
+	RID default_rd_samplers[RS::CANVAS_ITEM_TEXTURE_FILTER_MAX][RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX];
+	RID custom_rd_samplers[RS::CANVAS_ITEM_TEXTURE_FILTER_MAX][RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX];
+
 	/* GLOBAL VARIABLE API */
 
 	GlobalVariables global_variables;
@@ -221,6 +226,19 @@ public:
 
 	MaterialStorage();
 	virtual ~MaterialStorage();
+
+	/* Samplers */
+
+	_FORCE_INLINE_ RID sampler_rd_get_default(RS::CanvasItemTextureFilter p_filter, RS::CanvasItemTextureRepeat p_repeat) {
+		return default_rd_samplers[p_filter][p_repeat];
+	}
+	_FORCE_INLINE_ RID sampler_rd_get_custom(RS::CanvasItemTextureFilter p_filter, RS::CanvasItemTextureRepeat p_repeat) {
+		return custom_rd_samplers[p_filter][p_repeat];
+	}
+
+	void sampler_rd_configure_custom(float mipmap_bias);
+
+	// void sampler_rd_set_default(float p_mipmap_bias);
 
 	/* GLOBAL VARIABLE API */
